# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Leaderboard Service
  version: 1.0.0
  description: >-
    Seasonal game leaderboard. Design and rationale: docs/design.md.
    This file is maintained by hand and kept in sync with src/api/app.ts.
    Worker metrics are not part of this API: the outbox worker exposes
    GET /metrics on its own port (9100 by default), see docs/design.md section 3.4.
paths:
  /score:
    post:
      summary: Add points to a player in the active season
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [player_id, score_delta]
              properties:
                player_id: { $ref: '#/components/schemas/PlayerId' }
                score_delta:
                  type: integer
                  description: Non-zero. A zero delta would still move the player among equals.
                  minimum: -2097151
                  maximum: 2097151
                  not: { const: 0 }
                  example: 100
      responses:
        '200':
          description: New accumulated score (source of truth)
          content:
            application/json:
              schema:
                type: object
                required: [player_id, season_id, score]
                properties:
                  player_id: { $ref: '#/components/schemas/PlayerId' }
                  season_id: { type: integer, example: 202609 }
                  score: { type: integer, example: 1500 }
        '400': { $ref: '#/components/responses/ValidationError' }
        '422':
          description: Resulting score is out of the supported range
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: score_out_of_range }
        '503':
          description: No active season, or PostgreSQL is unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                no_active_season: { value: { error: no_active_season } }
                storage_unavailable: { value: { error: storage_unavailable } }
  /leaderboard/top:
    get:
      summary: Top-N players of the active season
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        '200':
          description: Top players, ordered by rank
          content:
            application/json:
              schema:
                type: object
                required: [season_id, entries]
                properties:
                  season_id: { type: integer }
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/Entry' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /leaderboard/rank/{player_id}:
    get:
      summary: Player position and neighbours
      parameters:
        - name: player_id
          in: path
          required: true
          schema: { $ref: '#/components/schemas/PlayerId' }
        - name: n
          in: query
          description: Number of neighbours above and below
          schema: { type: integer, minimum: 0, maximum: 50, default: 5 }
      responses:
        '200':
          description: Player and neighbours
          content:
            application/json:
              schema:
                type: object
                required: [season_id, player, above, below]
                properties:
                  season_id: { type: integer }
                  player: { $ref: '#/components/schemas/Entry' }
                  above:
                    type: array
                    description: Higher-ranked neighbours, best first
                    items: { $ref: '#/components/schemas/Entry' }
                  below:
                    type: array
                    items: { $ref: '#/components/schemas/Entry' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404':
          description: Player is not in the leaderboard of the active season
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: player_not_ranked }
        '503': { $ref: '#/components/responses/Unavailable' }
  /health:
    get:
      summary: Liveness probe
      description: >-
        The process is up and serves HTTP. Does not check PostgreSQL or Redis, so a
        storage outage does not make the orchestrator restart the API.
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Liveness' }
              example: { status: ok }
  /ready:
    get:
      summary: Readiness probe
      description: Checks PostgreSQL and Redis; the pod should receive traffic only on 200.
      responses:
        '200':
          description: PostgreSQL and Redis both answered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Readiness' }
              example: { status: ok, postgres: true, redis: true }
        '503':
          description: At least one storage did not answer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Readiness' }
              example: { status: degraded, postgres: true, redis: false }
components:
  schemas:
    PlayerId:
      type: string
      minLength: 1
      maxLength: 128
      example: player-42
    Entry:
      type: object
      required: [rank, player_id, score]
      properties:
        rank: { type: integer, minimum: 1, example: 1 }
        player_id: { $ref: '#/components/schemas/PlayerId' }
        score: { type: integer, example: 1500 }
    Liveness:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
    Readiness:
      type: object
      required: [status, postgres, redis]
      properties:
        status: { type: string, enum: [ok, degraded] }
        postgres: { type: boolean }
        redis: { type: boolean }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - no_active_season
            - score_out_of_range
            - storage_unavailable
            - leaderboard_unavailable
            - player_not_ranked
    ValidationError:
      description: Produced by the Fastify JSON schema validator, not by the handlers.
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode: { type: integer, example: 400 }
        error: { type: string, example: Bad Request }
        message: { type: string, example: body/score_delta must be integer }
  responses:
    ValidationError:
      description: Request body, query or path parameter failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
    Unavailable:
      description: >-
        No active season; Redis is unavailable; or PostgreSQL is unavailable and the
        active season has never been loaded (otherwise the last known season is used)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            no_active_season: { value: { error: no_active_season } }
            storage_unavailable: { value: { error: storage_unavailable } }
            leaderboard_unavailable: { value: { error: leaderboard_unavailable } }